megabanx-invoicing-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.3"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.49"}
asyncpg = "^0.31.0"
alembic = "^1.18.4"
python-dotenv = "^1.2.2"
jinja2 = "^3.1.6"
weasyprint = "^68.1"
aiosmtplib = "^5.1.0"
python-multipart = "^0.0.24"
pillow = "^12.2.0"
aiofiles = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
